% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_longitudinal.R
\name{model_longitudinal}
\alias{model_longitudinal}
\alias{model_longitudinal_linear}
\alias{model_longitudinal_itp}
\alias{model_longitudinal_idp}
\title{Model Creation: Longitudinal Models}
\usage{
model_longitudinal_linear(mu_a, sigma_a, t_max)

model_longitudinal_itp(mu_a, sigma_a, a_c1 = 0, b_c1 = 1, t_max)

model_longitudinal_idp(
  mu_a,
  sigma_a,
  a_c1 = 0,
  b_c1 = 1,
  a_c2 = -1,
  b_c2 = 0,
  t_max
)
}
\arguments{
\item{mu_a, sigma_a, a_c1, b_c1, a_c2, b_c2}{hyperparameters of the specified
longitudinal model.  See below for parameterization.}

\item{t_max}{a scalar, typically indicating the latest observed time
for subjects.  This will influence the interpretation of the
parameters of each model.}
}
\description{
Model Creation: Longitudinal Models
}
\section{Longitudinal Linear}{

  Let \eqn{f(d)} be a dose response model.  The expected value of the
  response, y, is:
  \deqn{E(y) = g(d, t)}
  \deqn{g(d, t) = a + (t / t_max) * f(d)}
  \deqn{a \sim N(mu_a, sigma_a)}
}

\section{Longitudinal ITP}{

  Let \eqn{f(d)} be a dose response model.  The expected value of the
  response, y, is:
  \deqn{E(y) = g(d, t)}
  \deqn{g(d, t) = a + f(d) * ((1 - exp(- c1 * t))/(1 - exp(- c1 * t_max)))}
  \deqn{a \sim N(mu_a, sigma_a)}
  \deqn{c1 \sim Uniform(a_c1, b_c1)}
}

\section{Longitudinal IDP}{

  Increasing-Decreasing-Plateau (IDP).

  Let \eqn{f(d)} be a dose response model.  The expected value of the
  response, y, is:
  \deqn{E(y) = g(d, t)}
  \deqn{g(d, t) = a + f(d) * (((1 - exp(- c1 * t))/(1 - exp(- c1 * d1))) *
    I(t < d1) + (1 - gam * ((1 - exp(- c2 * (t - d1))) /
    (1 - exp(- c2 * (d2 - d1))))) *
    I(d1 <= t <= d2) + (1 - gam) * I(t > d2))}
  \deqn{a \sim N(mu_a, sigma_a)}
  \deqn{c1 \sim Uniform(a_c1, b_c1)}
  \deqn{c2 \sim Uniform(a_c2, b_c2)}
  \deqn{d1 \sim Uniform(0, t_max)}
  \deqn{d2 \sim Uniform(d1, t_max)}
  \deqn{gam \sim Uniform(0, 1)}
}

